Fix spelling 'CronEventSchedulerVertical'.
